Added backend and connection to MongoDB
- Implemented a backend using Node and Express
- Connected backend to MongoDB
- Currently storing the number of times each search button is clicked
