+++ frontend/package.json
@@ -1,4 +1,5 @@
 {
+  "proxy": "http://localhost:5000",
   "name": "pocketlegalbeta3",
   "version": "0.1.0",
   "private": true,
@@ -10,6 +11,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.5",
     "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
